Providing inversion information

CrossInverts.AbstractCrossInversionCase — Type

AbstractCrossInversionCase

Interface for providing all relevant information for a cross-individual mixed effects bayesian inversion.

Concrete types should implement

get_priors_dict(::AbstractCrossInversionCase, indiv_id; scenario) Priors for model parameters in fixed, random, and indiv effects.

get_priors_random_dict(::AbstractCrossInversionCase; scenario) Priors for meta-parameters for random effects.

get_obs_uncertainty_dist_type(::AbstractCrossInversionCase; scenario) Type of distribution of observation-uncertainty per stream.

get_indivdata(::AbstractCrossInversionCase, indiv_id; scenario) The times, observations, and uncertainty parameters per indiv_id and stream.

CrossInverts.get_priors_dict — Function

get_priors_dict(::AbstractCrossInversionCase, indiv_id; scenario)

Provide a dictionary (par -> Distribution) for prior parameters and unknowns.

CrossInverts.get_priors_random_dict — Function

get_priors_random_dict(::AbstractCrossInversionCase; scenario)

Provide a dictionary (par -> Distribution) for prior parameters and unknowns.

CrossInverts.get_obs_uncertainty_dist_type — Function

get_obs_uncertainty_dist_type(::AbstractCrossInversionCase; scenario)

Provide the type of distribution of observation uncertainty for given stream, to be used with fit_mean_Σ.

CrossInverts.get_indivdata — Function

get_indivdata(::AbstractCrossInversionCase, indiv_id; scenario)

Provide Tuple (indiv_id -> (stream_info) for each individ. Where StreamInfo is a Tuple `(streamsymbol -> (;t, obs, obstrue)). Such that solution can be indexed by sol[streamsymbol](t) to provide observations. Valueobs_true` is optional. They are synthetic data without noise generated from the system, which are not used in inversion, but used for comparison.

The ValueType dispatches to different implementations. There is am implementation for Val(:CrossInverts_samplesystem1) independent of scenario.

CrossInverts.df_from_paramsModeUpperRows — Function

df_from_paramsModeUpperRows(paramsModeUpperRows)

Convert Tuple-Rows of (:par, :dType, :med, :upper) to DataFrame. And Fit distribution.

Setup tools

CrossInverts.get_indiv_parameters_from_priors — Function

get_indiv_parameters_from_priors(inv_case::AbstractCrossInversionCase,
        indiv_ids, mixed_keys;
        scenario, system,
        rng = StableRNG(234),
        priors_dict = get_priors_dict(inv_case, missing; scenario),
        priors_random_dict = get_priors_random_dict(inv_case; scenario),)

Construct a DataFrame with parameters across sites with

fixed parameters corresponding to the mean of its prior

random parameters corresponding to mean modified by sampled random effects and its meta parameters

individual parameters

Value

DataFrame with columns indiv_id, u0 and p, with all states and parameters of the given system as ComponentVector labelled by get_state_labeled and get_par_labeled.

CrossInverts.setup_tools_mixed — Function

setuptoolsmixed(pindiv::DataFrame; invcase, scenario = NTuple{0, Symbol}(), system, mixedkeys, psets = setuppsetsmixed(invcase; scenario, mixed_keys, system))

Given inv_case, the keys for different mixed effects, individual state (u0, p given with p_indiv), and individual priors, sets up NamedTuple of

mixed: mixed effects NamedTuple(fixed, random, indiv, indiv_random) from individual's states and parameters

df: DataFrame p_indiv extended by columns

paropt: optimized parameters extracted from indiviudals state and parameters

tools: tools initialized for each site (see setup_tools_indiv)

psets: NTuple{ODEProblemParSetter} for each mixed component

priors_pop: ComponentVector of priors on population level (fixed, random, random_σ)

sample0: ComponentVector of an initial sample This can be used to name (attach_axis) a sample from MCMCChains object

Simulating

CrossInverts.gen_sim_sols_probs — Function

sim_sols_probs(fixed, random, indiv, indiv_random)

Update and simulate system (given with tools to gensimsols_probs) by

for each individual i

update fixed parameters: fixed

update random parameters: random .* indiv_random[:,i]

update indiv_id parameters: indiv[:,i]

simulate the problem

return a vector(nindiv) of (;sol, problemopt)

If non-optimized p and u0 differ between individuals, they must already be set in tools[i_indiv].problem.

CrossInverts.gen_model_cross — Function

function gen_model_cross(;
        inv_case::AbstractCrossInversionCase, tools, priors_pop, sim_sols_probs,
        scenario, psets, solver)

Generate a Turing model using objects generated by setup_tools and forward simulation by gen_sim_sols_probs.